SimpleEvolve

Documentation for SimpleEvolve.

SimpleEvolve.MultiChannelSignal

SimpleEvolve.SignalSUM

SimpleEvolve.TanhEnvelope

SimpleEvolve._step

SimpleEvolve.amplitude_ti

SimpleEvolve.amplitude_ws

SimpleEvolve.costfunction_direct_exponentiation

SimpleEvolve.costfunction_ode_ssvqe

SimpleEvolve.evolve_ODE_multiple_states

SimpleEvolve.infidelity

SimpleEvolve.penalty_gradient

SimpleEvolve.projector

SimpleEvolve.reconstruct

SimpleEvolve.reconstruct_gradient_ti

SimpleEvolve.single_step

SimpleEvolve.MultiChannelSignal — Type

MultiChannelSignal 
A struct to represent a multi-channel signal with multiple digitized signals.

SimpleEvolve.SignalSUM — Type

SignalSum{T<:AbstractFloat}

Algebraic sum of multiple signal components.

SimpleEvolve.TanhEnvelope — Type

TanhEnvelope{T<:AbstractFloat}

Signal with hyperbolic tangent envelope for smooth transitions.

Fields

amplitude::T: Peak amplitude

sigma::T: Rise/fall time constant

center::T: Central time position

SimpleEvolve._step — Function

Auxiliary function to evolve a single step in time.

SimpleEvolve.amplitude_ti — Method

amplitude_ti(signal::DigitizedSignal, t) Calculate the amplitude of a signal at time t using trigonometric interpolation Args: signal: Input digitized signal t: Time at which to evaluate the signal Returns: Amplitude of the signal at time t

SimpleEvolve.amplitude_ws — Method

amplitudews(signal::DigitizedSignal, t; windowradius=8) Calculate the amplitude of a signal at time t using Whittaker-Shannon interpolation from the signal samples of reduced size Args: signal: Input digitized signal t: Time at which to evaluate the signal window_radius: Number of samples on each side of the center sample (default is 8) Returns: Amplitude of the signal at time t

SimpleEvolve.costfunction_direct_exponentiation — Method

costfunctiondirectexponentiation(ψ0, eigvals, signal, nsites, drives, T, ntrottersteps, Costham)

cost_function for the system using direct exponentiation
args:
    ψ0     : Initial state
    eigvals: Eigenvalues of the Hamiltonian
    signal : Signal to be evolved
    n_sites: Number of sites in the system
    drives : External drives applied to the system # annhilation operators in case of qubits
    T      : Total time for evolution
    n_trotter_steps: Number of Trotter steps for exponentiation
    Cost_ham: Hamiltonian used for cost function calculation
returns:
    cost   : Cost function value
    ψ_direct: Evolved state

SimpleEvolve.costfunction_ode_ssvqe — Method

costfunctionodeexcitedstates(ψ0, eigvals, signal, nsites, drives, eigvectors, T, Costham; basis="eigenbasis", tolode=1e-8)

cost_function for the system using ODE solver for excited states
args:
    ψ0     : Initial state
    eigvals: Eigenvalues of the Hamiltonian
    signal : Signal to be evolved
    n_sites: Number of sites in the system
    drives : External drives applied to the system # annhilation operators in case of qubits
    eigvectors: Eigenvectors of the Hamiltonian
    T      : Total time for evolution
    Cost_ham: Hamiltonian used for cost function calculation
    basis  : Basis in which the evolution is performed (default is "eigenbasis")
    tol_ode: Tolerance for ODE solver (default is 1e-8)
returns:
    cost   : Cost function value
    ψ_ode  : Evolved state

costfunction=sumi real(ψi' * Costham * ψ_i)

SimpleEvolve.evolve_ODE_multiple_states — Method

function to evolve multiple states using ODE solver simultaneously

SimpleEvolve.infidelity — Method

infidelity(ψ, φ) Function to compute the infidelity between two state vectors args: ψ : First state vector φ : Second state vector returns: infidelity value

SimpleEvolve.penalty_gradient — Method

penalty_gradient(Ω::Vector{ComplexF64}, Ω₀::Float64) Computes the gradient of the penalty term for the control amplitudes Ω. df(x)/dx = sign(x) * h * (1 + 1 / y^2) / Ω₀ for y > 0, y= abs(x) - 1 where h = exp(y - 1 / y)

Args:
    Ω   : Control amplitudes
    Ω₀  : Reference amplitude
Returns:
    grad : Computed gradient of the penalty term

SimpleEvolve.projector — Method

projector(n_sites::Integer, n_levels::Integer, n_levels0::Integer)

Project a Hilbert space of n_sites n_levels0-level qubits onto that of n_sites n_levels-level qubits

Returns an (n_sites^n_levels, n_sites^n_levels) shaped matrix Π. To perform the projection on a vector, use ψ ← Πψ. To perform the projection on a matrix, use A ← ΠAΠ'.

SimpleEvolve.reconstruct — Method

reconstruct(signal::DigitizedSignal, outputsamples::Int, method::Symbol; secondarymethod=:polynomial, weight=0.5, polyorder=4, windowradius=8, filtercutoffratio=0.8) Reconstruct a signal using specified interpolation methods. Args: signal: Input digitized signal output_samples: Desired output length method: Interpolation method to use Returns: Reconstructed signal with anti-aliasing

SimpleEvolve.reconstruct_gradient_ti — Method

reconstructgradientti(signal::DigitizedSignal, output_samples::Int)

Reconstruct a signal using trigonometric interpolation.
Args:
    signal: Input digitized signal
    output_samples: Desired output length
Returns:
    Reconstructed signal with anti-aliasing

SimpleEvolve.single_step — Function

Auxiliary function to evolve a single step in time.