SimpleEvolve

Documentation for SimpleEvolve.

SimpleEvolve.MultiChannelSignal

SimpleEvolve.amplitude_ti

SimpleEvolve.amplitude_ws

SimpleEvolve.costfunction_direct_exponentiation

SimpleEvolve.infidelity

SimpleEvolve.projector

SimpleEvolve.reconstruct

SimpleEvolve.reconstruct_gradient_ti

SimpleEvolve.MultiChannelSignal — Type

MultiChannelSignal 
A struct to represent a multi-channel signal with multiple digitized signals.

SimpleEvolve.amplitude_ti — Method

amplitude_ti(signal::DigitizedSignal, t) Calculate the amplitude of a signal at time t using trigonometric interpolation Args: signal: Input digitized signal t: Time at which to evaluate the signal Returns: Amplitude of the signal at time t

SimpleEvolve.amplitude_ws — Method

amplitudews(signal::DigitizedSignal, t; windowradius=8) Calculate the amplitude of a signal at time t using Whittaker-Shannon interpolation from the signal samples of reduced size Args: signal: Input digitized signal t: Time at which to evaluate the signal window_radius: Number of samples on each side of the center sample (default is 8) Returns: Amplitude of the signal at time t

SimpleEvolve.costfunction_direct_exponentiation — Method

costfunctiondirectexponentiation(ψ0, eigvals, signal, nsites, drives, T, ntrottersteps, Costham)

cost_function for the system using direct exponentiation
args:
    ψ0     : Initial state
    eigvals: Eigenvalues of the Hamiltonian
    signal : Signal to be evolved
    n_sites: Number of sites in the system
    drives : External drives applied to the system # annhilation operators in case of qubits
    T      : Total time for evolution
    n_trotter_steps: Number of Trotter steps for exponentiation
    Cost_ham: Hamiltonian used for cost function calculation
returns:
    cost   : Cost function value
    ψ_direct: Evolved state

SimpleEvolve.infidelity — Method

infidelity(ψ, φ) Function to compute the infidelity between two state vectors args: ψ : First state vector φ : Second state vector returns: infidelity value

SimpleEvolve.projector — Method

projector(n_sites::Integer, n_levels::Integer, n_levels0::Integer)

Project a Hilbert space of n_sites n_levels0-level qubits onto that of n_sites n_levels-level qubits

Returns an (n_sites^n_levels, n_sites^n_levels) shaped matrix Π. To perform the projection on a vector, use ψ ← Πψ. To perform the projection on a matrix, use A ← ΠAΠ'.

SimpleEvolve.reconstruct — Method

reconstruct(signal::DigitizedSignal, outputsamples::Int, method::Symbol; secondarymethod=:polynomial, weight=0.5, polyorder=4, windowradius=8, filtercutoffratio=0.8) Reconstruct a signal using specified interpolation methods. Args: signal: Input digitized signal output_samples: Desired output length method: Interpolation method to use Returns: Reconstructed signal with anti-aliasing

SimpleEvolve.reconstruct_gradient_ti — Method

reconstructgradientti(signal::DigitizedSignal, output_samples::Int)

Reconstruct a signal using trigonometric interpolation.
Args:
    signal: Input digitized signal
    output_samples: Desired output length
Returns:
    Reconstructed signal with anti-aliasing